Catppuccin wallpapers, these are weird
Some of these are symlinked into the gruvbox directory, these are
hardcoded to `/home/emily/dotfiles/catppuccin-wallpapers`. I'll try to
fix this at some point but for now this is the best way I could find to
do this. Remove all symlinks in the directory if you don't want them
crashing your random-wallpaper script